+++ module-apps/application-desktop/windows/Names.hpp
@@ -13,4 +13,5 @@ namespace app::window::name
     inline constexpr auto desktop_pin_lock    = "PinLockWindow";
     inline constexpr auto desktop_locked      = "LockedInfoWindow";
     inline constexpr auto desktop_update      = "Update";
+    inline constexpr auto desktop_mmi_pull    = "MmiPullWindow";
 }; // namespace app::window::name